refactor: move event-handlers.js to game-logic
