+++ src/game-logic/event-handlers.test.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 
-import Farmhand from './Farmhand'
-import { stageFocusType, fieldMode } from './enums'
-import { testItem } from './test-utils'
+import Farmhand from '../Farmhand'
+import { stageFocusType, fieldMode } from '../enums'
+import { testItem } from '../test-utils'
 
-jest.mock('./data/items')
-jest.mock('./data/levels', () => ({
+jest.mock('../data/items')
+jest.mock('../data/levels', () => ({
   levels: [
     {
       id: 0,
@@ -18,8 +18,8 @@ jest.mock('./data/levels', () => ({
   ],
   itemUnlockLevels: {},
 }))
-jest.mock('./data/recipes')
-jest.mock('./data/shop-inventory')
+jest.mock('../data/recipes')
+jest.mock('../data/shop-inventory')
 
 let component
